\name{ctgt-package}
\alias{ctgt-package}
\alias{ctgt}
\docType{package}
\title{
  'Closed Testing with Globaltest for Pathway Analysis'
}
\description{
 A shortcut procedure for closed testing with the global test is presented. 
}
\details{
  See examples in \link{actgt} function.
}
\author{
Ningning Xu

Maintainer: Ningning Xu <n.xu@lumc.nl; xu15263142750@gmail.com>
}
\references{
  Ningning Xu, Aldo solari, Jelle Goeman, Clsoed testing with global test, with applications on metabolomics data, arXiv:2001.01541, https://arxiv.org/abs/2001.01541
  Jelle J. Goeman, Sara A. van de Geer, Floor de Kort, Hans C. van Houwelingen, A global test for groups of genes: testing association with a clinical outcome, Bioinformatics, Volume 20, Issue 1, 1 January 2004, Pages 93-99, https://doi.org/10.1093/bioinformatics/btg382
}
\keyword{ package }