\name{majorizing vector}
\alias{getL}

\title{The majorizing vector}

\description{To get the majorizing vector at a specific level, given the upbound and lowbound.}

\usage{
  getL (ub, lb, level)

}

\arguments{
  \item{ub}{upper bound.}
  \item{lb}{lower bound.}
  \item{level}{level of interest.}
  
  
}

\value{Returns a numeric vector with the same length as ub and lb. }

\author{
Ningning Xu

Maintainer: Ningning Xu <n.xu@lumc.nl; xu15263142750@gmail.com>
}
\references{
   Ningning Xu, Aldo solari, Jelle Goeman, Clsoed testing with global test, with applications on metabolomics data, arXiv:2001.01541, https://arxiv.org/abs/2001.01541
}

